fix: StringVector use Utf8Array (#222)

+++ src/datatypes/src/arrow_array.rs
@@ -10,8 +10,8 @@ use crate::value::Value;
 
 pub type BinaryArray = ArrowBinaryArray<i64>;
 pub type MutableBinaryArray = ArrowMutableBinaryArray<i64>;
-pub type MutableStringArray = MutableUtf8Array<i64>;
-pub type StringArray = Utf8Array<i64>;
+pub type MutableStringArray = MutableUtf8Array<i32>;
+pub type StringArray = Utf8Array<i32>;
 
 macro_rules! cast_array {
     ($arr: ident, $CastType: ty) => {
